requirements.txt
sympy
pyqrcode
pypng
pyzbar
pillow
mysql-connector-python
django
flask
schedule